criu/files-reg.c: move file validation functions
This moves the file validation helper function declarations to
files-validation.h and the function definitions to files-validation.c.
It also updates Makefile.crtools.
