+++ criu/Makefile.crtools
@@ -25,6 +25,7 @@ obj-y			+= file-ids.o
 obj-y			+= file-lock.o
 obj-y			+= files-ext.o
 obj-y			+= files.o
+obj-y			+= files-validation.o
 obj-y			+= files-reg.o
 obj-y			+= fsnotify.o
 obj-y			+= image-desc.o
